+++ src/NzbDrone.Core.Test/DecisionEngineTests/ClusterAnalysis/SortedClusterResultFixture.cs
@@ -1,11 +1,10 @@
-﻿using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using FluentAssertions;
 using NUnit.Framework;
-using NzbDrone.Core.DecisionEngine.Clustering;
+using NzbDrone.Core.DecisionEngine.ClusterAnalysis;
 
-namespace NzbDrone.Core.Test.DecisionEngineTests.Clustering
+namespace NzbDrone.Core.Test.DecisionEngineTests.ClusterAnalysis
 {
     [TestFixture]
     public class SortedClusterResultFixture
